requirements.txt
mysql-connector-python
langchain
langchain-openai
python-dotenv

